+++ src/jtl/Connector/Presta/Controller/Product.php
@@ -1,12 +1,20 @@
 <?php
 namespace jtl\Connector\Presta\Controller;
 
+use Attribute;
+use AttributeGroup;
+use Combination;
+use Context;
+use Exception;
+use jtl\Connector\Model\Identity;
 use jtl\Connector\Presta\Utils\Utils;
+use PrestaShopDatabaseException;
+use PrestaShopException;
 
 class Product extends BaseController
 {
     private static $idCache = array();
-
+    
     public function pullData($data, $model, $limit = null)
 	{
 		$limit = $limit < 25 ? $limit : 25;
@@ -26,7 +34,8 @@ class Product extends BaseController
 			$model = $this->mapper->toHost($data);
 
 			$return[] = $model;
-
+            $this->pullSpecialAttributes($data, $model);
+			
 			$count++;
 		}
 
@@ -41,7 +50,7 @@ class Product extends BaseController
 
 			foreach ($resultVars as $data) {
 				$model = $this->mapper->toHost($data);
-
+				
 				$return[] = $model;
 			}
 		}
@@ -66,8 +75,14 @@ class Product extends BaseController
 
         \Product::flushPriceCache();
     }
-
-	public function pushData($data)
+    
+    /**
+     * @param \jtl\Connector\Model\Product $data
+     * @return mixed
+     * @throws PrestaShopDatabaseException
+     * @throws PrestaShopException
+     */
+    public function pushData($data)
 	{
         if (isset(static::$idCache[$data->getMasterProductId()->getHost()])) {
             $data->getMasterProductId()->setEndpoint(static::$idCache[$data->getMasterProductId()->getHost()]);
@@ -83,7 +98,7 @@ class Product extends BaseController
 
             $id = $product->id;
         } else {
-            list($productId, $combiId) = explode('_', $data->getId()->getEndpoint());
+            list($productId, $combiId) = array_pad(explode('_', $data->getId()->getEndpoint(), 2), 2, null);
 
             $product = new \Product($masterId);
 
@@ -126,10 +141,11 @@ class Product extends BaseController
                 $id = $data->getMasterProductId()->getEndpoint().'_'.$combiId;
             }
 
-            $combi = new \Combination($combiId);
+            $combi = new Combination($combiId);
 
             $valIds = array();
-
+            $attrGrpId = null;
+            
             foreach ($data->getVariations() as $variation) {
                 $attrNames = array();
                 foreach ($variation->getI18ns() as $varI18n) {
@@ -141,12 +157,12 @@ class Product extends BaseController
                         $attrNames[$langId] = $varName;
                     }
 
-                    if ($langId == \Context::getContext()->language->id) {
+                    if ($langId == Context::getContext()->language->id) {
                         $attrGrpId = $this->db->getValue('SELECT id_attribute_group FROM '._DB_PREFIX_.'attribute_group_lang WHERE name="'.$varName.'"');
                     }
                 }
 
-                $attrGrp = new \AttributeGroup($attrGrpId);
+                $attrGrp = new AttributeGroup($attrGrpId);
                 $attrGrp->name = $attrNames;
                 $attrGrp->public_name = $attrNames;
                 $attrGrp->group_type = 'select';
@@ -166,7 +182,7 @@ class Product extends BaseController
                             $valNames[$langId] = $valName;
                         }
 
-                        if ($langId == \Context::getContext()->language->id) {
+                        if ($langId == Context::getContext()->language->id) {
                             $valId = $this->db->getValue('
                               SELECT l.id_attribute
                               FROM '._DB_PREFIX_.'attribute_lang l
@@ -176,7 +192,7 @@ class Product extends BaseController
                         }
                     }
 
-                    $val = new \Attribute($valId);
+                    $val = new Attribute($valId);
                     $val->name = $valNames;
                     $val->id_attribute_group = $attrGrpId;
 
@@ -218,6 +234,7 @@ class Product extends BaseController
             if (isset($product) && $data->getMasterProductId()->getHost() === 0) {
                 ProductAttr::getInstance()->pushData($data, $product);
                 ProductSpecific::getInstance()->pushData($data, $product);
+                $this->pushSpecialAttributes($data, $product);
             }
         }
 
@@ -225,7 +242,14 @@ class Product extends BaseController
 
 		return $data;
 	}
-
+    
+    /**
+     * @param \jtl\Connector\Model\Product $data
+     * @return mixed
+     * @throws PrestaShopException
+     * @throws PrestaShopDatabaseException
+     * @throws Exception
+     */
     public function deleteData($data)
     {
         $isCombi = (strpos($data->getId()->getEndpoint(), '_') === false) ? false : true;
@@ -235,11 +259,11 @@ class Product extends BaseController
         } else {
             list($productId, $combiId) = explode('_', $data->getId()->getEndpoint());
 
-            $obj = new \Combination($combiId);
+            $obj = new Combination($combiId);
         }
 
         if (!$obj->delete()) {
-            throw new \Exception('Error deleting product with id: '.$data->getId()->getEndpoint());
+            throw new Exception('Error deleting product with id: '.$data->getId()->getEndpoint());
         }
 
         return $data;
@@ -263,4 +287,60 @@ class Product extends BaseController
 
         return ($count + $countVars);
 	}
+    
+    /**
+     * @param \jtl\Connector\Model\Product $data
+     * @param \Product $product
+     * @throws PrestaShopException
+     */
+    private function pushSpecialAttributes($data, $product)
+    {
+        $specialAttributes = ProductAttr::getSpecialAttributes();
+
+        $foundSpecialAttributes = [];
+        foreach ($data->getAttributes() as $attribute) {
+            foreach ($attribute->getI18ns() as $i18n) {
+                $name = array_search($i18n->getName(), $specialAttributes);
+                if($name === false) {
+                    $name = $i18n->getName();
+                }
+
+                if (isset($specialAttributes[$name]) && $i18n->getValue() !== "") {
+                    $foundSpecialAttributes[$name] = $i18n->getValue();
+                    break;
+                }
+            }
+        }
+        
+        foreach ($foundSpecialAttributes as $key => $value) {
+            if ($value === 'false' || $value === 'true') {
+                $value = filter_var($value,FILTER_VALIDATE_BOOLEAN);
+            } elseif (is_numeric($value)) {
+                $value = (int)$value;
+            }
+            
+            $product->{$specialAttributes[$key]} = $value;
+        }
+        
+        $product->save();
+    }
+    
+    /**
+     * @param $data
+     * @param \jtl\Connector\Model\Product $model
+     */
+    private function pullSpecialAttributes($data, $model) {
+        foreach (ProductAttr::getSpecialAttributes() as $wawiName => $prestaName) {
+            $attribute = new \jtl\Connector\Model\ProductAttr();
+            $attributeI18n = new \jtl\Connector\Model\ProductAttrI18n();
+            $attribute->setId(new Identity($prestaName));
+            $attribute->setProductId($model->getId());
+            $attributeI18n->setProductAttrId($attribute->getId());
+            $attributeI18n->setLanguageISO(Utils::getInstance()->getLanguageIsoById((string)Context::getContext()->language->id));
+            $attributeI18n->setName($wawiName);
+            $attributeI18n->setValue($data[$prestaName]);
+            $attribute->setI18ns([$attributeI18n]);
+            $model->addAttribute($attribute);
+        }
+    }
 }