+++ changelog.md
@@ -1,6 +1,8 @@
 1.5.9
 ------
 - CO-521 - Fixed a bug regarding cover images
+- CO-530 - Implemented product special attributes like "products_status"
+
 1.5.8.1
 ------
 - Fixed a bug on ProductAttr push